Introduce BaseSymbol class and improve docs for symbols
This fixes "flip" of arguments in relational expressions,
mentioned in the issue sympy/sympy#7951.

The docs says [1] (see also [2]) about relational ops:
There are no swapped-argument versions of these methods (to be used when
the left argument does not support the operation but the right argument
does); rather, __lt__() and __gt__() are each other’s reflection,
__le__() and __ge__() are each other’s reflection, and __eq__() and
__ne__() are their own reflection.  If the operands are of different
types, and right operand’s type is a direct or indirect subclass of the
left operand’s type, the reflected method of the right operand has
priority, otherwise the left operand’s method has priority.

So, if B is a subclass of A, we can't avoid calling the "reflected"
method.  For example, A() < B() will be silently transformed to
B() > A().  Only know workaround per se - avoid subclassing of "public"
classes.  For example, old class hierarchy for symbols is forbidden:

                    Symbol
                   /      \
                 Dummy    Wild

In this patch, common base class for symbols was implemented.

References
==========

.. [1] https://docs.python.org/3.5/reference/datamodel.html#object.__lt__
.. [2] https://bugs.python.org/issue22052
.. [3] https://github.com/sympy/sympy/issues/7951