Use srepr instead of sstr for __repr__ printing
As it stated by the Python docs [1] about returned value of
__repr__(), "If at all possible, this should look like a valid
Python expression that could be used to recreate an object
with the same value (given an appropriate environment)."

It seems, that our printing system violates this maxima.  "More
accurate rule - is that repr() output for SymPy's objects is something
that can be "a suitable input for sympify()" [2].   An example is
Rational's printing:

In [3]: repr(Rational(1, 2))
Out[3]: '1/2'

I believe, we should prefer Python's conventions.

This closes skirpichev/omg#32.

References
==========

.. [1] https://docs.python.org/3/reference/datamodel.html#object.__repr__
.. [2] https://github.com/sympy/sympy/issues/6880#issuecomment-37008096