dsolve: fix typo in code for lin_2eq_type2
Add a test which I though would not pass, but with a fix to a
trivial typo it does pass!  Fractional increase in my opinion of
this code.

At least one more case untested: added a FIXME.  I know this is
uncovered because of the typo mentioned above, neither branch
was getting tested.  (The "if" branch is tested by the test
added here).