Fix the test_issue_6828() test
Previously it was using .equal() to compare two expressions containing floating
point, which is incorrect --- .equal() is using symbolic manipulations to see
if the two expressions are equal, which cannot work for a floating point. The
only way to compare expressions that contain floating point is to substitute
random numbers for the symbols and making sure the (numerical) answer is close
enough to zero.

We use verify_numerically() to do this.