Added type checking to Transform
I added type checking to the transform code for Point and Point3D. I also added new tests to ensure that the changed code functions properly. Now before
a transformation is applied to the point we check that the supplied variable is a matrix and that it is of the correct dimension.