Separate metaclass hierarchy from class hierarchy
ManagedProperties would use BasicMeta both as metaclass and superclass,
so __init__ and __new__ are called both for class and for object
creation. Such a situation confuses maintainers, libraries, and tools
alike.
FunctionClass had a similar condition, it used BasicMeta as metaclass
and ManagedProperties as superclass; it is now firmly a subclass of
ManagedProperties, i.e. a metaclass.

Also, BasicType served no real purpose anymore, so it was removed,
leaving BasicMeta as the root of the metaclass hierarchy.

The only non-metaclass code that needed a concomitant change was
`sympify()`: Metaclasses are not a "standard sympifiable class"
anymore, which it was assuming in a test.

// edited by skirpichev