Added 'global_vars' option to codegen.
Previously, all symbols in a given expression were converted to function
arguments. However, sometimes the generated function is part of larger
module where global variables are defined and it is not desirable to
pass these in as arguments to the function.

With this commit the behavior is now:

 >>> from sympy.utilities.codegen import codegen
 >>> from sympy.abc import x, y, z
 >>> [(f_name, f_code), header] = codegen(
 ...     ("f", x+y*z), "F95", header=False, empty=False,
 ...     argument_sequence=(x, y), global_vars=(z,))
 >>> print(f_code)
 REAL*8 function f(x, y)
 implicit none
 REAL*8, intent(in) :: x
 REAL*8, intent(in) :: y
 f = x + y*z
 end function

This makes it much easier to generate uniform function signatures.

// edited by skirpichev