solvers: Added to the solve() docstring
Working the final example in the tutorial when solving for the radius of a
sphere in terms of volume I could not figure out how to alleviate unwanted
solutions.  Being a Mathematica man I assumed the assumptions would be an
argument to solve() but there was little mention of it in the docstring.
I myself feel that this should at least be mentioned, so I added it to the
note at the end of the Example section.

I would even consider adding explanations to the solve() flags.  A good example
of a docstring there that made no since to me was under force.

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>