geometry: Correct Polygon.encloses_point and update references
The Polygon.encloses_point method is based of the Python code at
http://www.ariel.com.au/a/python-point-int-poly.html, which is in turn
based off C code at
http://local.wasp.uwa.edu.au/~pbourke/geometry/insidepoly/. The second
link is now broken. It's new home is at
http://paulbourke.net/geometry/polygonmesh/#insidepoly. The code at the
new link shows two lines of the `for` loop in the Ariel and Sympy code
are improperly indented. This commit fixes the indentation and updates
the references.

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>