Cleanup Set.contains
Removing this line beacause the logic behind this line is
wrong, beacause no subclasses return None for _contain.
it seems that this should check for every finite set but
simpify return none for this line to be tested, but for finite
set simpify does not return none.so both statement are controductory
to each other.

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>