Shortcut cancelling fractions with integer one
Fractions with integer one in either the numerator or the denominator
would make it possible to cancelled out any common factor from the
numerator and the denominator.  However, in the current implementation,
there is no special treatment for this situation.  The computation of
the GCD with integer one would take a long time by using the Euclidean
algorithm, as in the issue sympy/sympy#9341.

As a result, in this commit, when called with expressions, the cancel
function will directly return the quotient when an integer unity is
found.  This could solve issue #9314.  And when the function is called
with polynomials directly or called with tuples, this short-cut will not
be invoked to better preserve its original behaviour in internal use.

Add one as special case in computing GCD of DMP

In the GCD computation of dense multivariate polynomials, previously in
function `sympy.polys.euclidtools._dmp_rr_trivial_gcd`, only the special
cases with one of the polynomials being zero is specially treated as
trivial case.  However, the polynomial of constant unity is the unit of
the multiplication operation in the ring, so it cannot possibly have
other GCD with any polynomial.  So now one is being handled as a special
case as well to skip the later expensive operations.

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>