Docs and code fixes for "Combinatorics" module
1. `.is_normal` for PermutationGroup with `strict`(optional)

Add `strict` argument for `.is_normal` method for the PermutationGroup,
similar to `strict` argument for `.is_subgroup` method, it also does the same
thing to make the degree of both the groups same, like
```
>>> s4 = SymmetricGroup(4)
>>> s6 = SymmetricGroup(6)
>>> s6.is_normal(s4, strict=False)
True
```

Earlier this used to raise IndexError, the default value of `strict` is True

2. PermutationGroup() with no argument contains only IdentityElm

Avoid the `ValueError` caused, since PermutationGroup() with no
arguments. One can think of the group generated by an empty
args list as the symmetric group of the empty set, that consists of the
identity element alone.

So set `args = [Permutation()]` and then to go on with no error message.
It should rather return `SymmetricGroup(0)` after PR #10363 is complete.

// edited by skirpichev

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>