Reintroduce bounded/unbounded assumptions
For sets (or Symbol's, that represent sets) - these mean
bounded/unbounded sets.  But for Symbol's, that represent
numbers - usual finite/infinite assumption.

We don't support yet set algebra (i.e. expressions with Interval's), but
this seems to be a good feature in future.  But already right now we should
keep in mind that things like sin(oo) are not numbers and finite or
extended_real assumptions - not valid for them.

The bounded/unbounded pair of assumptions - could be helpful in such
cases as a generalization of the finite/infinite pair for sets.  Please
note, that right now it's used only in the heuristics() function
of the limit module.