Allow PolynomialRing's without generators
The standard methods for the construction of polynomials, .as_poly
and Poly, are able to find the minimal coefficient domain for a
polynomial given as an expression. However, they can only be used
in case there is at least one independent generator because the
representation they create, the 'dense' polynomial, is a nested list
of lists whose depth is the number of generators. Otherwise, a dummy
temporary generator has to be added for the construction of the
coefficient domain.

The representation of 'sparse' polynomials is different. It is a
dictionary whose keys are monomials of the polynomial ring.
The monomials are tuples of integers, one for each generator. The
empty tuple can be used to represent the unit monomial of
a polynomial ring with no generators.

This PR implements sparse polynomial rings with no generators. They are
not intended for practical use, only for convenient construction of
coefficient domains. In fact, such polynomial rings are isomorphic
to their coefficient domains, and therefore not needed for other
purposes.

The construction of both dense and sparse polynomials makes use of the
same routines at the lowest level. They currently raise a
GeneratorsNeeded exception if no generators are found. To allow
the construction of sparse polynomials with no generators, these
routines have been modified to return the tuple of generators even
if there are none. An exception is raised only when a dense polynomial
construction is attempted.

// edited by skirpichev

Based on sympy/sympy#11589

We use instead a different convention for degree's (empty
degree_list() and None for degree/tail_degree.

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>