Don't split integrals that can't be computed
Issue sympy/sympy#2708 was (partially) fixed in commit
6fe73a1f by evaluating separately the terms of the antiderivative.
That gave rise to another issue sympy/sympy#10445, which was
fixed in sympy/sympy#10515 by removing the separation.

This commit aims to combine the merits of the two PRs by treating
separately those terms of the antiderivative that contain unevaluated
integral factors and evaluating the rest together in one sum.

// edited by skirpichev

Also handle case of empty factors in dup_factor_list()

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>