Use instances of Mapping rather than dict in subs
Closes sympy/sympy#10710.  All standard library `dict`-like objects
are instances of `Mapping` thus, `OrderedDict`, `ChainMap`, etc., can now
be a drop-in replacement for a `dict` in substitutions.  Rather than
having to wrap the object in a `dict` call.

In Python 3.3 `collections.abc` separated, however to keep compatibility
with `sys.version_info < (3,3)` `from collections import Mapping` works
for newer and older versions of Python.

// edited by skirpichev

Dict now registered to Mapping type.

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>