Implemented function on symbol can give ValueError
Closes sympy/sympy#10810

Sympy may try and simplify functions such as f(0.), by evaluating them
numerically.  When doing this, the floating point value (such as 0.) has
been converted to a sympy Float object.  Trying to evaluate the symbolic
float argument with numerical implemented functions can cause a variety
of errors.  Around line 468 of core/function.py:eval_evalf, there is a
check for these kinds of errors, falling back to the not-evaluated form.
At the moment eval_evalf only allows implemented functions to return
TypeError and AttributeError, otherwise the routine raises the error
and crashes out.

I recently came across a case where a numerical function
(scipy.interpolate.interp1d) returns a ValueError if we pass it symbols
to operate on, which seems a reasonable thing to do, and something that
may be true of other implemented functions receiving unexpected sympy
arguments to evaluate.

Allow ValueError as well as the other valid errors.

// edited by skirpichev

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>