Replace the coefficient of abs in the lookup table of meijerint.
The denominator of the coefficient  pi/(cos(pi*a/2)*gamma(a))
becomes indeterminate for negative odd integers  a  where  cos(pi*a/2)
has a zero and  gamma(a)  has a pole. There is no code for computing
the limit at such argument values. This PR replaces the coefficient
with  2*sin(pi*a/2)*gamma(1 - a), which is well defined for  re(a) < 1,
since the only poles of gamma occur at nonpositive integers.

The equivalence of the two forms can be derived from the standard
formulas

    gamma(a)*gamma(1 - a) = pi/sin(pi*a)

and

    sin(pi*a) = 2*sin(pi*a/2)*cos(pi*a/2).

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>