Change the default value for the modules lambdify argument to "numpy"
["math", "mpmath", "sympy"] is still used if NumPy is not installed.

Previously, if NumPy was installed, it used
["math", "numpy", "mpmath", "sympy"], but this lead to confusing behavior
where NumPy would be used only if the function wasn't in the math module.
Since the primary use-case of lambdify is to convert SymPy expressions to
NumPy expressions, and since NumPy arrays only work if used with NumPy
functions, this default makes more sense.

We should perhaps also change the non-numpy default, although mixing and
matching math and mpmath functions isn't as big of a deal (at least for float
input).

This fixes sympy/sympy#2866.

This is technically an API change, so some discussion is needed if this should
have a deprecation cycle.

// edited by skirpichev

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>