Check hyper() b-parameters in test_hyperexpand
If any of the b-parameters of a hypergeometric series is a negative
integer or zero, the series is not well defined unless there is an
a-parameter of the same type with smaller absolute value, in which
case the series is a polynomial.

In this patch, the the b-parameters are checked only if there are
randomly generated parameters to allow test_polynomial pass.
(a-parameters could also be checked in case there is a bad b-parameter,
but it is easier to just generate a set of new ones since there are
separate tests for polynomial cases.)

Fixes sympy/sympy#11881

// edited by skirpichev

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>