elliptic_integrals: rewrite to use standard notation for z, m
The incomplete elliptic integrals depend on "the amplitude" z and "the
parameter" m.  The complete elliptic integrals depend only on the
parameter, taking the value pi/2 for the amplitude.  In some cases,
"z" was used for the parameter both in the code and in the docs.
The latter can be confusing b/c it breaks from standard notation (and
especially b/c Elliptic integrals have various different common conventions
in use).

This commit changes the docs as well as the code.  It makes no functional
changes.

Based on sympy/sympy#12137

Signed-off-by: Sergey B Kirpichev <skirpichev@gmail.com>