UndefinedFunction should not evalf
Suppose we call `Function('sin')` to create an UndefinedFunction.  It has
a suspcious name but is nevertheless still an undefined function and so
it should not evalf.  Add tests.  Fixes sympy/sympy#6938.

The fix is to return early in the `_eval_evalf` if the `func` is an
`UndefinedFunction`, instead of looking through mpmath tables for a match etc.
Thanks to @smichr for this hint.

Note I first tried to fix this by introducing `UndefinedFunction._eval_evalf`.
The reason that does not work is that `f(x)` is not itself an
`UndefinedFunction` (only `f` is).

// edited by skirpichev