polys: fix sorting of real roots
In the triggered by solve_poly_inequality() example, the problem
polynomial is something like (x + 1)**3*(3*x + 1).  In this case, for
distinct roots, we have isolating intervals like r1=[-1, -1] (this root is
from the ground domain, ZZ) and r2=[-1, 0] (for x=-1/3).  Second root is not
going to be improved by sorting helper, since r1.b==r2.a.

Closes sympy/sympy#20902