Cylindrical algebraic decomposition implementation
This is an implementation of cylindrical algebraic decomposition (CAD)
in Diofant, which enables the solution of systems of multivariate
polynomial (in)equalities. Example:

    >>> from diofant.abc import x,y
    >>> from diofant.solvers.polysys import solve_poly_system_cad
    >>> solve_poly_system_cad([-x**2 - 1 > 0], [x])
    []
    >>> solve_poly_system_cad([x*y > 2, 10*x < 1, x**2 + y > 3], [x, y])
    [{y: -2, x: -sqrt(5) - 1}]

The two main functions that expose this functionality are (1)
cylindrical_algebraic_decomposition(), which computes a CAD adapted to a
set of polynomials (a decomposition into cells over which each
polynomial is sign-invariant); and (2) solve_poly_system_cad(), which
takes a set of polynomial (in)equalities, constructs polynomial atoms
(by subtracting the right-hand side), then passing it into
cylindrical_algebraic_decomposition() and lastly checking each cell for
the system to hold.

Technically, this required the implementation of: computing reducta
(red() and red_set()), computing subresultant coefficients
(subresultant_polynomials() and subresultant_coefficients()), and the
Hong projector operator (projone(), projtwo(), and hongproj()). With
these functions, implementing CAD is rather straightforward.

// edited by skirpichev

Mostly mechanical port of sympy/sympy#26785, pylint/flake8
errors are fixed.